LayoutTests/platform/glib/fast/xsl/xslt-bad-import-uri-expected.txt

CONSOLE MESSAGE: Document is empty

This tests that a bad @import URI does not cause a crash.
